IFilterFactory => IFilterProvider
